refactor: split DomainEntityMapper
