+++ src/main/java/com/apipietunes/clients/model/dto/MusicTrackDto.java
@@ -1,11 +1,11 @@
-package com.apipietunes.clients.model.dto.domain;
+package com.apipietunes.clients.model.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import com.apipietunes.clients.model.dto.domain.inner.InnerAlbumDto;
-import com.apipietunes.clients.model.dto.domain.inner.InnerBandDto;
+import com.apipietunes.clients.mapper.util.InnerAlbumDto;
+import com.apipietunes.clients.mapper.util.InnerBandDto;
 
 import java.util.UUID;
