+++ src/main/java/com/apipietunes/clients/model/dto/TrackLoaderResponse.java
@@ -1,8 +1,6 @@
 package com.apipietunes.clients.model.dto;
 
 
-import com.apipietunes.clients.model.dto.domain.MusicTrackDto;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
