+++ src/main/resources/application.properties
@@ -24,7 +24,7 @@ spring.servlet.multipart.max-request-size=${MULTIPART_MAX_REQUEST_SIZE}
 
 spring.main.allow-circular-references=true
 
- logging.level.org.springframework.data.neo4j=DEBUG
+# logging.level.org.springframework.data.neo4j=DEBUG
 
 # logging.level.root=TRACE
